refactor(stream): trim redundant client contract fields and test fixtures (#895)
