+++ frontend/store/__tests__/chatRuntime.interrupts.test.ts
@@ -17,118 +17,10 @@ import type {
   RuntimeStatusContract,
 } from "./chatRuntime.test.support";
 
-const normalizeRuntimeStateToken = (state: string) => {
-  const normalized = state.trim().toLowerCase().replace(/_/g, "-");
-  return normalized.startsWith("task-state-")
-    ? normalized.slice("task-state-".length)
-    : normalized;
-};
-
-const buildCanonicalInterrupt = (interrupt?: Record<string, unknown>) => {
-  if (!interrupt) {
-    return null;
-  }
-  const phase = typeof interrupt.phase === "string" ? interrupt.phase : "asked";
-  return {
-    ...interrupt,
-    phase,
-    source: "stream",
-  };
-};
-
-const buildStatusUpdate = ({
-  state,
-  seq,
-  messageId,
-  completionPhase,
-  interrupt,
-}: {
-  state: string;
-  seq?: number;
-  messageId?: string;
-  completionPhase?: string;
-  interrupt?: Record<string, unknown>;
-}) => ({
-  statusUpdate: {
-    status: { state },
-    metadata: {
-      shared: {
-        ...(interrupt ? { interrupt } : {}),
-        stream: {
-          ...(seq !== undefined ? { seq } : {}),
-          ...(messageId ? { messageId } : {}),
-          ...(completionPhase ? { completionPhase } : {}),
-        },
-      },
-    },
-  },
-  version: "v1",
-  runtimeStatus: {
-    state: normalizeRuntimeStateToken(state),
-    isFinal:
-      state === "TASK_STATE_COMPLETED" ||
-      state === "TASK_STATE_FAILED" ||
-      state === "TASK_STATE_INPUT_REQUIRED",
-    ...(buildCanonicalInterrupt(interrupt)
-      ? { interrupt: buildCanonicalInterrupt(interrupt) }
-      : {}),
-    ...(seq !== undefined ? { seq } : {}),
-    ...(completionPhase ? { completionPhase } : {}),
-    ...(messageId ? { messageId } : {}),
-  },
-});
-
-const buildArtifactUpdate = ({
-  agentMessageId,
-  eventId,
-  seq,
-  text,
-  source = "assistant_text",
-}: {
-  agentMessageId: string;
-  eventId: string;
-  seq: number;
-  text: string;
-  source?: string;
-}) => ({
-  artifactUpdate: {
-    op: "append",
-    artifact: {
-      artifactId: `${agentMessageId}:stream:${seq}`,
-      parts: [{ text }],
-      metadata: {
-        shared: {
-          stream: {
-            blockType: "text",
-            source,
-            messageId: agentMessageId,
-            eventId,
-            seq,
-          },
-        },
-      },
-    },
-  },
-  version: "v1",
-  streamBlock: {
-    eventId,
-    eventIdSource: "upstream",
-    messageIdSource: "upstream",
-    seq,
-    taskId: agentMessageId,
-    artifactId: `${agentMessageId}:stream:${seq}`,
-    blockId: `${agentMessageId}:primary_text`,
-    laneId: "primary_text",
-    blockType: "text",
-    op: "append",
-    source,
-    messageId: agentMessageId,
-    role: "agent",
-    delta: text,
-    append: true,
-    done: false,
-  },
-});
+import {
+  buildStatusUpdatePayload as buildStatusUpdate,
+  buildTextArtifactUpdatePayload as buildArtifactUpdate,
+} from "@/test-utils/streamContractFixtures";
 
 describe("executeChatRuntime empty-content recovery", () => {
   beforeEach(() => {
